osc out switched to TCP

+++ src/O_OSC/O_osc.h
@@ -18,7 +18,7 @@ public:
     void sendHigh();
     void sendFlash(string parameter);
     
-    void sendEncoder(string parameter, int message);
+    void sendEncoder(string parameter, float message);
     void fineEncoder(int message);
     void sendEncoderPercent(string parameter, int message);
